regsim: import the Linux kernel's regulatory.h
This starts the work of a regulatory simulator. We start off
by importing the Linux kernel's regulatory.h which is permissively
licensed. We also use a copy of the kernel's Signed-off-by
documentation so we can keep that practice of respecting each
file's licenses, although this whole project will be permissively
licensed.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>