regsim: simplify channel testing output
This simplifies the channel testing output by only listing
the frequency once and listing the allowed frequencies
bandwidths. The 40 MHz bandwidth highlights an issue in
design with the current framework: bandwidths that rely on
MIMO require categorizing the channel by type, either
HT40 Plus or HT40 Minus, that is the secondary channel
depends on which side of the center of freq its on.

This will be our next fix.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>