regsim: display current max EIRP for each bandwidth
Additionally throw in the respective IEEE channel number,
we take this helper from iw and provide proper attributions.

The next step is to consider the case where we actually want
to specify a higher EIRP than what the first rule found will
provide. Consider a set of rules that allow us to transmit
at max 20 dBm on channel 4 when using 20 MHz but if we use
the same channel and a bandwidth of 5 MHz we can transmit
at 30 dBm. In that case currently we'll only be able to
transmit at 20 dBm. We want to:

I.  Provide a desired target output power
II. Look for TPC specification, which may buy us max 3 dB

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>