regsim: move channel definition to regulatory.h
This lets us make regulatory.h independent of the wireless core.
The idea behind this is to enable a set of regulatory library helpers
to be to used independently of the wireless core. We will then have
a wireless core with a regulatory core that uses a regulatory library:

|-----------------------------------
| wireless core |  regulatory core |
------------------------------------
                         |
                         |
                         |
                         v
               ----------------------
               | regulatory library |
               ----------------------

With this split up we can define helper data structures within
the regulatory library that the wireless core and regulatory can
use independently of the functional aspects of the regulatory
library.

The regulatory library can then be used in userspace or kernelspace
with its own set of control framework for event management / locking
without changing. The library also enables uses of the same code in
any OS with little or no porting.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>