regsim: update a device's channels based on set regdomain
This adds support for updating a device's channel based on the
set regulatory domain. This will currently terminate with a BUG_ON()
which we have defined for userspace as abort() given that we do not yet
have the regcore->last_request set. We will need to address how we define
each new regulatory request next and see how we split this up between the
reglib.c and the user code on reg.c.

Worth noting is that the regsim code already has support for target power
and bandwidth lookups on the regulatory data but to truly support this we'd
need to add a reglib call to enable wireless core code to update the target
EIRP on the channel whenver power gets changed. This would ensure that
when we need to do a new lookup for regulatory for the device, we'll be
optimizing the configuration for the device dependening on each channel's
target power / badnwidth. In other words this would be neede to takes
advantage of the new overlapping regulatory rule support on regsim.

It may be worth noting that when we update the power for a channel we likely
also want to update the rule for only that channel, this will have to be
implemented on regsim.c but consideration must be made to the cost of this
implementation -- it means we'd do an O(n) lookup on the regulatory rules for
whenever we update the target power. If we cache this for a range of target
EIRPs we'd have closer to an O(1) lookup at the cost of expanding the channel
data structure with more data.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>