regsim: fix race condition when starting core and registering devices
Thers is a small race condition between which we register the wireless
core and allow devices to register themselves to the core. The race
is that the wireless core can be registered without the first
regulatory hint request being processed given that this is offloaded
onto a queue and processed later but we still allow devices to register
to the wireless core. Devices should still technically be allowed to
register to the wireless core even if we have not processed the first
regulatory hint given that we the first regulatory hint is for an updated
world regulatory domain and we already have a static world regulatory
domain available to us upon initialization -- the first regulatory hint
simply asks for an updated world regulatory domain, should one exist.

Fix this by making the last request always be understood to be the
static world regulatory domain.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>