regsim: add regulatory_exit()
All the work we did before to help terminiting was added
to add the code for regulatory_exit(). In particular I noticed
that we were not waiting for the regulatory hint to complete even
if the thread was spawned. With this code one could simulate a
regulatory hint that in-kernel takes a while to process as follows:

	--- a/reg.c
	+++ b/reg.c
	@@ -21,8 +21,10 @@ static DECLARE_WORK(reg_work, reg_todo);
	  * This lets us keep regulatory code which is updated on a regulatory
	  * basis in userspace.
	  */
	+#include <unistd.h>
	 static int call_crda(const char *alpha2)
	 {
	+       sleep(3);
		if (!reglib_is_world_regdom((char *) alpha2))
			printf("Calling CRDA for country: %c%c\n",
			       alpha2[0], alpha2[1]);

What you expect to see is for this thread to finish running
before regulatory_exit() completes. This is exactly what the
cancel_work_sync() does for us.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>