regsim: add communication simulation module
This adds a communication module to the simulator in charge
of acting as the interface between the simulated kernel and
userspace. It'd behave as a respective OS kernel's ioctl / nl80211
message parser. We keep this very simple and simply run the code
that we need to run for userspace within the program instead of
splitting the two.

Up next is to actually run the CRDA code to process the request
and hand it off to reg.c

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>